requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
